requirements.txt
discord.py
youtube_dl
python-dotenv
